startup: fix logic/design error that led to deep errors
The old state in this commit would lead to an existing session
being loaded with the claim that it was a new session. This went
unnoticed until 4bed642d7198, where newness impacted the time
domain of the session (and actually led to it being set to a random memory
value).