Unload VST3 modules when closing session
This is perhaps a better solution than b8551eed7ef
and 8d0a6556080 and 7942897d93136. It is certainly less
fragile.

It is more consistent with other plugin standards,
where modules are closed with the last instance in a session.

Then again keeping the VST3 factory around is beneficial
when switching snapshots.

Discuss, and let's watch for issues when re-loading a
previously unloaded VST3 module.