Fix for deadlock adding MIDI track when using pipewire jack
Without this fix one thread is waiting in the reader lock for the `_io_lock`
mutex in `IO::connected_latency` (io.cc:1275), while another thread acquired
the mutex with a writer lock in `IO::ensure_ports` (io.cc:490). The deadlock
comes from the function `ensure_ports` being waiting in
`_session.engine().register_input_port(...)`, which is calling
`jack_port_register` and this one is locked inside `pw_thread_loop_wait`.

On the other hand, there is the thread from `JACKAudioBackend::process_thread`
that is calling `jack_cycle_wait` in an infinite loop which in turns calls
another pipewire function that polls.

Maybe this is not a deadlock, but a starvation problem because I don't
understand why `jack_port_register` doesn't make progress.

The solution consist of moving the writer lock used by `ensure_ports` inside
`ensure_ports_locked` and reduce its scope to the minimum. In particular,
the mutex is not held while calling `register_input_[in|out]put_port`.