Use a dedicated mutex for the fix in 5925ee0
Using the same mutex in AudioEngine::buffer_size_change and AudioEngine::stop
causes a deadlock at exit using pipewire-jack due to a lock inversion.
While stop() holds the _process_lock mutex the buffer size callback is called
while holding a mutex inside pipewire than stop needs to acquire, leading to a
deadlock.

An apparent solution is to use a dedicated mutex for the purpose for which the
lock in buffer_size_change was introduced: mutual exclusion between this function
and the processing loop.