osc: Remove unused OSCSelectObserver::nplug_params variable
This is only set to 0 and never read. In plugin_init(), a value is
calculated for this variable, but because it declares a local variable
with the same name, the value is never assigned. And since the value is
only ever used inside plugin_init(), there is no need to keep the
instance variable.

This prepares the code for sending feedback for multiple plugins instead
of just the selected one.