osc: Add feedback bit to send feedback about all plugins
Currently, feedback is only sent about parameter for a single, selected
plugin. By setting this new "all_plugins" bit in the `/set_surface`
message (or preferences), behaviour is changed to provide feedback for
all plugins in the currently selected strip.

If enabled, this bit changes the format of the `/select/plugin/*` messages
by adding an extra (first) argument containing the plugin id (piid).
This matches the format that is accepted by the
`/select/plugin/parameter` message already (which allows the piid to be
specified or omitted).

The code that tracks the selected plugin is still active even when the
all_plugins bit is set, but the current plugin is not used (i.e. the
normal messages without the piid are not sent anymore to prevent
confusion about the meaning of the messages).

Plugin (parameter) paging is still used and applies to all plugins at
the same time.

If the "send ssid as path" bit is set, this extra piid argument is also
put into the path. Since the existing `*_message_with_id` helper
functions only support a single id argument and we needed two, an extra
local helper function (`plugin_parameter_message`) is introduced that
conveniently handles all variants of these messages needed (ids as
argument or path, including a piid or not, including a paid or not, with
string, float or int value).

This results in a long function calls and duplicates a bit of code from
the `*_message_with_id` helper functions, but the alternative of adding
`*_message_with_two_ids` (or something) helper functions would probably
end up with more duplicate code and more complex calls, so this is
probably the cleanest approach.

This function also includes a usleep call which was copied from the
`*_message_with_id` helper functions verbatim. Without it, some messages
seem to be dropped. This has a bit of code smell (race condition?), but
that is out of scope for this commit.

To clear parameter values that are not longer present (i.e. when
switching to a plugin with fewer parameters), the existing `plug_size`
variable (that tracked the number of parameters sent on the previous
update) is generalized and renamed to a `params_sent` vector, that is
index by piid and contains the number of parameters sent for each
plugin. Element 0 is used for the selected plugin when the all_plugins
feedback bit is not set, or unused when the bit is set.