osc: Make list of plugin input parameter for all plugins
Previously, a list of input parameters was made by _sel_plugin for the
currently selected plugin. Now, it happens whenever the current strip is
changed for all plugins in the current strip. This prepares for supporting
plugin parameter feedback for all plugins instead of just a single one in an
upcoming commit.

This might result in a little less or a little more work, depending on the
usage pattern (changing strips is more work, changing plugins and back is less
work).