Fix allocation bug in PBD::StackAllocator
allocate and deallocate methods currently check if the pointer is between
&_buf and &_buf + stack_capacity, but the buffer is between &_buf and
&_buf + (sizeof (T) * stack_capacity). However, _ptr is then summed by the
amount of elements to allocate for, which means that it is actually working
as a counter of elements allocated in the function stack starting from the
value of the address of _buf. While this might look like a neat optimization
to avoid a multiplication, it introduces a bug. Subsequent allocations
return a pointer to the wrong memory address, allowing for the overwrite of
previously allocated elements.