Define _GNU_SOURCE in hidapi build
I tried compiling Ardour 8.7 on openSUSE Tumbleweed but the build failed due to problems with resolution of functions like `wcsdup`, `strtok_r`, etc. It seems that even if the correct headers are included, those functions won't be included unless `_GNU_SOURCE` is defined (the explanation is in `wchar.h`). 

I added the definition of `_GNU_SOURCE` to the `wscript` in the case of Linux builds so that the code will be compiled correctly.