audio_backend: Rename instantiate() parameter arg2 to `session_id`
The thing called backend_client_name defaulted to lowercase PROGRAM_NAME
and could be controlled with:

    -c, --name <name>           Use a specific backend client name, default is ardour

This value was *only* used for set_backend arg2 which only is used for
jack, where it was called session_uuid and passed to jack_client_open as
"a SessionID Token this allows the sessionmanager to identify the client
again" because JackSessionID.

It is certainly misleading to call it a uuid. It also doesn't seem
helpful to call it "client name" when the documentation calls it
"session id". And mostly: jack_client_open calls its *first* argument
"client_name" and it comes from argv1.

To clear things up, consistently call it something with "session id".
And it is not helpful to use a generic arg2 naming.