Fix crashes when moving mouse over port-matrix
Beware of using float as loop/while variable.

When moving the mouse over the tilted port-label area the position p
is set to ` p - (_height - o) * tan (angle ())`.
In optimized (fast math) builds this can be come negative (-0.0).

`::type_channel_to_overall()` then returns a negative channel count,
leading to a crash.

And in some edge-cases (literally moving the mouse to the edge after
resizing the window on Intel Mac), PortMatrixColumnLabels::motion is
called and the port corresponding to the mouse-position is also
computed to be negative.