SessionDialog: Use plain notebook tab labels for New/Recent/Open instead of custom ArdourButton
In my opinion, the ArdourButton approach for SessionDialog introduced in
524073d8 regressed in some annoying ways:
* No keyboard shortcuts.
* Big panes look a bit like Tcl/Tk.
* YELLING.

While it seems like keyboard shortcuts for ArdourButton would be handy
in many places, it doesn't seem trivial to add it.

Plain GtkNotebook seems to work better:
* Keyboard shortcuts for changing tabs.
* More standard look.

With these changes, New and Recent works pretty well with keyboard
navigation. There should only be minor changes when using mouse. The
buttons are of course smaller, but that not be a problem for users that
can navigate other parts of the Ardour UI.

Keyboard navigation in GtkFileChooser remains broken for now. There is
thus not much use of keyboard navigation to open the Open page, but we
add it anyway.