SessionDialog: Drop set_can_focus() for Gtk::CheckButton _disable_plugins
_disable_plugins was the only widget in SessionDialog that had
set_can_focus(). That made it look like something special was going on
there. But that is not the case. "can_focus" is already the default, and
there is no point in setting it again.

Remove it, and make it a bit more clear what is going on with focus
control.