Revert "avoid double-free of MIDI sys ex and patch change canvas items"
This reverts commit 0ddc05055523c5a014e6c3c630541be168f27ced.

It is not clear what this commit was really trying to do - when canvas items
are deleted they are removed from their parent, so just *not* deleting them
should not by itself fix an issue.

It turns out that not deleting the canvas item means that it remains in the
canvas scene graph to deliver events to, at which time it references its
now-deleted owner object, leading to a crash.

We need to keep an eye open for crashes resembling the double-free that this
reverted commit claimed to fix.