an initial fix for a crash occuring when deleting routes that lead to RouteGroup deletion
The RouteGroup containing the deleted routes emits the RouteRemoved signal; Session handles this
and if the RouteGroup is now empty, deletes the RouteGroup *while in the middle of the signal
emission process*. This deletes the Signal (since it was owned by the RouteGroup) which leads
to a crash inside the signal emission code, since that is now running "on" a deleted Signal
object.

This change simply makes that signal static, and so deleting the RouteGroup has no effect
on the Signal itself.

More changes to come related to better (shared_ptr<RouteGroup>) mgmt of RouteGroup lifetimes.