fix editing of meter & tempo points
when these points have times set in beat time (which they nearly always do), converting
their superclock time to samples and back loses accuracy. this caused the TempoMap to
find the wrong earlier points in the map.

It is still up for consideration that the editing code ought to identify explicitly
that we're working on an existing point, or adding a new one, and avoid lookup
by time for the existing points